refactor: extract switch echo filtering into SwitchEchoMixin (backlog 5.1)
cover_base.py had grown to 4776 lines. Move the self-contained "Switch echo
filtering" section — the relay echo-filter and wait_for_relay_feedback
machinery, 19 methods, ~535 lines — verbatim into a new SwitchEchoMixin in
cover_echo_filter.py, mirroring the existing CalibrationMixin / cover_calibration
pattern. cover_base.py drops to 4241 lines; CoverTimeBased gains the mixin in
its base list.

Behaviour-preserving: the method bodies move unchanged and the full suite passes
unchanged (1956). The only test changes are mock patch-target relocations — the
moved module globals (time, async_call_later, the RELAY_FEEDBACK_* constants)
now live in cover_echo_filter, so patches that targeted cover_base retarget
there; no assertion is changed. The two full-flow clock fixtures patch time on
both modules, since they drive moved and non-moved code.
